Failsafe for addModels receiving string/number
Some other target scripts like qTarget may allow adding models by string or integer one at a time instead of requiring a table of models to add. This will check the incoming data type and if it is not a table, it will make it a table to prevent errors.